Add saving images as stacks
Normally I imaged 20 per repeat and saving separate images was a
good idea, not the whole stack have to be read from the disk
in order to access one frame or so.

Wtb_sinusoidal recordings however are much longer and results huge
number of individual images, even though images are cropped and
therefore small in size. These huge amounts of small images are
slow to copy on NAS drive and access from there.
Stacks better.