Add static imaging (rotation triggered)
Previously "static imaging" was done by pupil imsoft sending a
physical trigger signal to the camera, which was initialised
in MicroManager and set to wait trigger. Imsoft wrote a file of
the rotations (1-to-1 order with the saved image filenames)

This is now a soft trigger, much slower. In addition, the images
are saved differently compared to the previous static imaging
(now the same way as the "dynamic imaging" images)